requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Core dependencies (auto-installed but pinned for safety)
pydantic==2.10.3
pydantic-settings==2.6.1

# AI
google-genai==1.16.0

# Database
psycopg2-binary==2.9.10

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Linting
ruff==0.8.4